fix #2674 - update micropayment - casting amount as double
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@15687 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ includes/external/micropayment/class.micropayment_helper.php
@@ -36,7 +36,7 @@ class micropayment_helper
             //'shop_version' => $this->getShopSignatur(),
             'currency'     => $order->info['currency'],
             'project'      => MODULE_PAYMENT_MCP_SERVICE_PROJECT_CODE,
-            'amount'       => (int)($order->info['pp_total'] * 100),
+            'amount'       => (double)($order->info['pp_total'] * 100),
             'orderid'      => $insert_id,
             'paytext'      => str_replace('#ORDER#',$insert_id,MODULE_PAYMENT_MCP_SERVICE_PAYTEXT),
             'theme'        => MODULE_PAYMENT_MCP_SERVICE_THEME,
